onesept

+++ libft/src/ft_strrchr.c
@@ -3,16 +3,16 @@
 /*                                                        :::      ::::::::   */
 /*   ft_strrchr.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: alsaeed <alsaeed@student.42abudhabi.ae>    +#+  +:+       +#+        */
+/*   By: alsaeed <alsaeed@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/01/10 15:31:45 by alsaeed           #+#    #+#             */
-/*   Updated: 2023/07/24 15:51:44 by alsaeed          ###   ########.fr       */
+/*   Updated: 2023/09/01 14:36:58 by alsaeed          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../inc/libft.h"
 
-char	*ft_strrchr(char *s, int c)
+char	*ft_strrchr(char *s, size_t c)
 {
 	size_t	len;
